Section 7 and 8 - Structure of Arrays, Indexing, Subsetting, Slicing and Iterating through Array

+++ day-15/python/section-7-structure-of-array.py
@@ -6,9 +6,12 @@ import numpy as np
 # itemsize: Memory used by each array element in bytes.
 # ndim: Number of axes(of dimensions).
 
-structuresOfArray = np.ones((5, 3))
+structuresOfArray = np.ones((5, 3), dtype=int)
 print("Sample Array: \n{}".format(structuresOfArray))
 print("Data Type: {}".format(structuresOfArray.dtype))
 print("Item Size: {}".format(structuresOfArray.itemsize))
 print("N Dimention: {}".format(structuresOfArray.ndim))
 print("Array Shape: {}".format(structuresOfArray.shape))
+
+# Creating a 3-D array. Difficult to print it.
+# reshape() simply reshape a 1-D array.
